Remove failure if new unhealthy nodes exist on start
It's happened to me fairly often that a TF apply has timed-out because a node was unhealthy for some external reason.  I fix the external reason, and just want bouncer to continue waiting for the new instance to come up and then continue.  However, currently, it bails out immediately when you start it in the fashion, so I have to wait for that one instance to become healthy, and only then kick-off the TF apply to re-run.

I'm not sure why we need this precondition, as without it, bouncer will just immediately start-up in a "waiting for instance to become healthy" state, which seems just fine.