NewContext should actually be new
It shouldn't be based on the previous context, and therefore inherit its deadline and just try to implement an additional one - this will never work because the old deadline will always be the initial timeout time after launch, and we aren't actually renewing a timer at all. By just creating a brand new context every time instead and throwing the old one away, we're actually getting a fresh timer.